Increase cohesion and decouple
Breaking this into a library and a main package.
