+++ cmd/blink/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"os"
+
+	"github.com/lelandbatey/blink/cmd/cli"
+)
+
+func main() {
+	os.Exit(cli.Run())
+}
